Move `MessageFormatter` and `MessageReader` to `kafka.common`
