+++ core/src/main/scala/kafka/common/MessageFormatter.scala
@@ -1,0 +1,39 @@
+/**
+  * Licensed to the Apache Software Foundation (ASF) under one or more
+  * contributor license agreements.  See the NOTICE file distributed with
+  * this work for additional information regarding copyright ownership.
+  * The ASF licenses this file to You under the Apache License, Version 2.0
+  * (the "License"); you may not use this file except in compliance with
+  * the License.  You may obtain a copy of the License at
+  *
+  *    http://www.apache.org/licenses/LICENSE-2.0
+  *
+  * Unless required by applicable law or agreed to in writing, software
+  * distributed under the License is distributed on an "AS IS" BASIS,
+  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  * See the License for the specific language governing permissions and
+  * limitations under the License.
+  */
+
+package kafka.common
+
+import java.io.PrintStream
+import java.util.Properties
+
+import org.apache.kafka.clients.consumer.ConsumerRecord
+
+/**
+  * Typical implementations of this interface convert a `ConsumerRecord` into a type that can then be passed to
+  * a `PrintStream`.
+  *
+  * This is used by the `ConsoleConsumer`.
+  */
+trait MessageFormatter {
+
+  def init(props: Properties) {}
+
+  def writeTo(consumerRecord: ConsumerRecord[Array[Byte], Array[Byte]], output: PrintStream): Unit
+
+  def close() {}
+
+}
